changes have addressed the comments in the PR
